elasticsearch init script
Adds a script to initialize the mappings of the elasticsearch db
used by the api.

The location of the elasticsearch db server can be provided using
command-line arguments or guessed by the script looking at the
freezer-api configuration file.

The script asks confirmation prior to deleting each of the
elasticsearch types whose mapping needs to be updated,
but it can be instructed to always assume a confirmation to ease
the invocation by scripts or deployment tools.

Change-Id: I0938c7cfde9bae440bf85f536a56079429d772a9
Implements: blueprint freezer-api-db-init