update type deletion policy in db-init
Types with not-matching mappings are not deleted by default.
When elasticsearch fails to merge the new mapping with the
existing one an error is returned.

The type is deleted only when explicitly requested, either
interactively or providing the following command-line parameter:

    --erase

Adds the possibility to select the specific mapping to upload
with the following command-line parameter:

 --mapping <specific-mapping>

Change-Id: If18fdba770790d8af03475d45da28c2e40fb7da6