Return an error for negative limits and offsets
This changes the behavior of all of the collection-based API endpoints
so that they return a bad request error (HTTP 400) for limits less
than or equal to zero and offsets less than zero. The API
speicification has also been updated to reflect the new rules.

Additionally, this adds new integration tests to verify the added
functionality. To support these, the API client 'get_' functions
were amended to accept '**params' for all endpoints (not only
'get_backups').

Closes-Bug: 1592046
Change-Id: I9968b093c22c48082635c4c97f94bc66dcdfcb10