Use http_proxy_to_wsgi middleware
This sets up the HTTPProxyToWSGI middleware in front of Freezer.
The purpose of this middleware is to set up the request URL
correctly in case there is a proxy (For instance, a loadbalancer
such as HAProxy) in front of Freezer.

The HTTPProxyToWSGI is off by default and needs to be enabled
via a configuration value.

Depends-On: Iffd38a325204a3ec7380a7a56061866477d3d06e
Change-Id: I44d60863eefeb52891474653aa9fcf1ba57d50a1
Closes-bug: #1590608