Allow bootstrapping against clusters and unify host options
The current settings inside freezer-manage reflect on using host and
port which are no longer used.  In addition, there seems to be no
splitting of the values in freezer-manage which means that it would
create a problem when trying to install against a list of hosts as
the ElasticSearch driver will refuse to accept the list.

This patch unifies all references of to 'hosts' and changes the option
to use ListOpt which is comma seperated values in the configuration
file which are then split into an array by oslo.config, which avoids
having to split things inside freezer.

Closes-Bug: #1681931

Change-Id: I613d0672ec2b186cd0d6365e795b35d9cff26dcd