update pylint to 1.9.2
The older version of pylint being used does not work correctly under
python 3. In order to be able to update the pylint job to run under
python 3, we need to update the tool.

This patch updates to the latest version at this time. It also updates
and pins astroid, which was previously capped. Using a pin instead of
a cap should let us avoid issues with new releases while being
specific about which version to actually use.

This patch also fix lower-constraints.

Co-Authored-By: Nguyen Hai <nguyentrihai93@gmail.com>
Change-Id: I6443247ab110dc9d5e71d23dc40ceabb439e23be
Signed-off-by: Doug Hellmann <doug@doughellmann.com>