import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Mark freezer-api-ubuntu-xenial job as non-voting, it fails consistently
on this branch.

Change-Id: Ide5d63094367685ffaa5175ce86d5360dd7fbcfd
Story: #2002586
Task: #24296