Download, and install elasticsearch server for freezer-api
In devstack/pkg, there is an elasticsearch.sh for downloading,
installing, starting and uninstalling of elasticsearch server,
but it is not flexible. In the future, freezer-api needs to support
ubuntu-xenial and ubuntu-bonic, which requires different elasticsearch
server versions. In addition, the programming interface of elastic
search server varies greatly before and after 5.0.0, which requires
different code of freezer-api to support. Pre-5.0.0 version
of elasticsearch server cannot be installed on ubuntu-bonic.
For the convenience of modification, freezer-api maintains
the elastic search.sh script by itself.

Change-Id: Ie069e2fd1e29daf53d8b1490f152cd820df036fe