Check client capabilities on job creation
Scheduling backups of Cinder volumes using
centralized instance(s) of freezer-scheduler
requires executing tasks on behalf of the user.

Client capabilities provide a mechanism to
restrict which actions, modes, storages, and
engines can be assigned to freezer-scheduler
for security and workload control.

The core implementation of client capabilities
resides in freezer-scheduler.

The primary purpose of checking capabilities on
the freezer-api side is to prevent the assignment
of actions that cannot be executed by a specific
scheduler instance.

Implements: blueprint centralized-scheduler
Depends-On: I15a91e8232d07e5bbc2cc95a72b0405c61c64a72
Change-Id: Ie774f9e66ac373a43ef935775de3b18539c5bebd