Add wsgi module to freezer-api
Changes in python packaging tooling mean that the wsgi_scripts
functionality via PBR may not longer function.

We migrate freezer to use wsgi modules to ensure consistent behavior
and achieve the community goal.

For more details see the proposed OpenStack goal [1].

We also fix compatability with Falcon 4.1 which is a requirements for
2026.1

[1] https://review.opendev.org/c/openstack/governance/+/902807

Change-Id: I6f73ee5fe6194d2a5390dfb5389e336719332cb1
Signed-off-by: Dmitriy Rabotyagov <noonedeadpunk@gmail.com>